Rename 'service_discovery' package to 'discovery'
