+++ openvpn/discovery/dto/bandwidth.go
