+++ openvpn/discovery/factory.go
@@ -1,10 +1,10 @@
-package service_discovery
+package discovery
 
 import (
 	"github.com/mysterium/node/datasize"
 	"github.com/mysterium/node/identity"
 	"github.com/mysterium/node/money"
-	"github.com/mysterium/node/openvpn/service_discovery/dto"
+	"github.com/mysterium/node/openvpn/discovery/dto"
 	dto_discovery "github.com/mysterium/node/service_discovery/dto"
 	"time"
 )
