+++ openvpn/discovery/factory_test.go
@@ -1,9 +1,9 @@
-package service_discovery
+package discovery
 
 import (
 	"github.com/mysterium/node/identity"
 	"github.com/mysterium/node/money"
-	"github.com/mysterium/node/openvpn/service_discovery/dto"
+	"github.com/mysterium/node/openvpn/discovery/dto"
 	dto_discovery "github.com/mysterium/node/service_discovery/dto"
 	"github.com/stretchr/testify/assert"
 	"testing"
